Include shell escaping to avoid globbing files named "weka_*"
The current script invokes a shell command from python and fails
to properly escape a "weka_*" argument to $(find). This results
in files matching the "weka_*" in the user's homedir being
expanded into arguments to $(find), which either
 a) creates an invalid find command, or
 b) fails to remove the cgroup container

b) is much more involved to fix.
While the right way to solve this is probably to use python's
os.walk or similar, this is a quick fix to quote the argument.