Avoid lower-casing NIC and drive names - WEKAPP-307881
It's totally valid for NIC names to have uppercase letters in the
name - it's more common in the case of bonded NIC devices - e.g.
an admin might create "bond0-100G". Therefore we should not cast
every name to lowercase.
Similar reason for drives; there's nothing preventing a drive
from having an uppercase name, even though it's not conventional.