Add support for IMDSv2 in AWS - WEKAPP-305278
If IMDSv2 is in use, then the "standard" URL to check if we're in
AWS (http://169.254.169.254/latest/meta-data/) will return an error
and thus we think we're not in AWS.
If instead we test the retrieval of a token, this works for both
IMDSv2 and IMDSv1 - i.e. it's a more reliable test.