Limit text-replacement to JSON keys only - VT-57
The previous version replaced the JSON string

[{container: {compute0},  ip: [..., "ip-172-31-3-122.ec2.internal", ]}]
[{container: {drives0},   ip: [..., "ip-172-31-3-122.ec2.internal", ]}]
[{container: {frontend0}, ip: [..., "ip-172-16-1-2.ec2.internal", ]}]

with the replacement

['[{"container": "compute0",  "ip": [..., ""ip"-172-16-1-2.ec2.internal"]}]',
 '[{"container": "drives0",   "ip": [..., ""ip"-172-16-1-2.ec2.internal"]}]',
 '[{"container": "frontend0", "ip": [..., ""ip"-172-16-1-2.ec2.internal"]}]']

(Outputs have been edited for clarity and brevity).

Note the quoting of the "ip" key but ALSO the hostname component "ip-172.."

This fix simply limits it to the JSON key by including the colon, as that's
not going to appear in a hostname.