VT-184: Add checking to ensure that CPUs support x86_64-v3 instruction set
As per recent core base compiler changes, older Sandy Bridge or Ivy Bridge
CPUs are no longer supported. We need at least Haswell's x86_64-v3 options.

This code checks for the implied existence of x86_64-v3 by checking for
the existence of avx2 in /proc/cpuinfo as a reasonable proxy for the
presence of the complete instruction set.