Cst jack support checking kernel args (#427)
* Minor typo fixes

* Set known_issues_file to be a global variable

We're going to rely on per-version known issues more and more; it
makes sense to have a known global file rather than passing it around
between functions with a static string hidden deep in a function call.

* Support checking backends for known-problematic kernel versions

In 4.4.7.89 we started seeing iommu=pt cause problems with
non-Broadcom NICs. We should therefore start checking backends for
that argument.
This code adds the framework for checking kernel arguments on backends.

* Add in the known issue for iommu=pt on 4.4.7.89

WEKAPP-513581 saw a problem get introduced with iommu=pt. This is
the known issues reference for that version.